+++ CHANGELOG.md
@@ -22,6 +22,7 @@
   Common user tenant invitation path changed to `/account/tenant/{tenant}/invite` (#501, v25.13-alpha5)
 
 ### Fixes
+- Fix websocket authentication introspection (#513, v25.26-alpha12)
 - Remove a- prefix from resource list API parameters (#511, v25.26-alpha10)
 - Fix endpoint for obtaining global roles (#506, v25.26-alpha7)
 - Fix PKCE strength evaluation (#504, v25.26-alpha6)